Refactorizes the project focusing on the administrator page
This is the first commit of a series of commits refactorizing the code of the
application. In this first commit a general refactorization removing unused
elements, fixing minor errors and clearing the source code was done.

In addition, the source code related with the administrator page was carefully
reviewed and refactorized, fixing some important errors.