Fixes security errors
Admin-On-Rest do not redirect paths that do not match with any resource.
When this happens, the main layout is showed without any resource in the
side menu.

As this application was not using a "Dashboard" component, which is
located at the root path, when a "not-logged" user accessed the root
path, the main layout without any resource was shown.

This commit solves this problem by adding a "Dashboard" component with
basic information about the application.

In addition, the AuthController now rejects users with "PATIENT" role.